requirements.txt
pandas
openpyxl
python-dotenv
from_root
PyYAML
requests
-e .